+++ libs/graphs/include/mrpt/graphs/CMRVisualizer.h
@@ -48,11 +48,11 @@ class CMRVisualizer
 	~CMRVisualizer() override;
 	void drawNodePoints(
 		mrpt::opengl::CSetOfObjects::Ptr& object,
-		const mrpt::system::TParametersDouble* viz_params =
+		const mrpt::containers::yaml* viz_params =
 			nullptr) const override;
 	void drawEdges(
 		mrpt::opengl::CSetOfObjects::Ptr& object,
-		const mrpt::system::TParametersDouble* viz_params =
+		const mrpt::containers::yaml* viz_params =
 			nullptr) const override;
 
    private:
@@ -81,10 +81,10 @@ class CMRVisualizer<
 	~CMRVisualizer();
 	void drawNodePoints(
 		mrpt::opengl::CSetOfObjects::Ptr& object,
-		const mrpt::system::TParametersDouble* viz_params = nullptr) const;
+		const mrpt::containers::yaml* viz_params = nullptr) const;
 	void drawEdges(
 		mrpt::opengl::CSetOfObjects::Ptr& object,
-		const mrpt::system::TParametersDouble* viz_params = nullptr) const;
+		const mrpt::containers::yaml* viz_params = nullptr) const;
 
    private:
 };